Remove FAQ comment inconsistent with new log-in requirements
Trying to start a meeting on meet.jit.si without logging in now leads to the
"Waiting for a moderator..." dialog. The remark that "(everyone is a moderator
if you are using meet.jit.si)" becomes confusing in light of this. As a quick
fix, I've just deleted the confusing phrase.